requirements.txt
flask
flask-restx
python-dotenv
pyodbc
openai
requests